feat add missing query methods to immutable

+++ src/immutable/float/query/within.rs
@@ -1,8 +1,10 @@
 use crate::distance_metric::DistanceMetric;
 use crate::float::kdtree::Axis;
+use crate::float_leaf_simd::leaf_node::BestFromDists;
 use crate::immutable::float::kdtree::ImmutableKdTree;
 use crate::nearest_neighbour::NearestNeighbour;
 use crate::types::Content;
+use az::Cast;
 
 use crate::generate_immutable_within;
 
@@ -57,7 +59,7 @@ impl<
 use memmap::MmapOptions;
 
 let mmap = unsafe { MmapOptions::new().map(&File::open(\"./examples/immutable-doctest-tree.rkyv\").unwrap()).unwrap() };
-let tree = unsafe { rkyv::archived_root::<ImmutableKdTree<f64, u32, 3, 32>>(&mmap) };"
+let tree = unsafe { rkyv::archived_root::<ImmutableKdTree<f64, u64, 3, 32>>(&mmap) };"
     );
 }
